Remove internal conan usage for cmake test
Using conan as a package means it gets included as a -isystem header.
This will prevent warnings from the library files from being emitted. To
get those back conan packages cannot be used in this way.

+++ include/libembeddedhal/full_scale.hpp
@@ -146,7 +146,7 @@ constexpr static T increase_bit_depth(U p_value)
   T result = static_cast<T>(p_value) << width_difference;
 
   if (p_value > 0) {
-    for (int i = final_source_width; i < final_width_difference;
+    for (size_t i = final_source_width; i < final_width_difference;
          i += final_source_width) {
       result |= (result >> i);
     }
